Creating pre-compiled scripts android-application-setup.gradle.kts
