+++ build.gradle.kts
@@ -1,5 +1,4 @@
-import com.android.build.gradle.BaseExtension
-import com.android.build.gradle.BasePlugin
+
 import org.gradle.api.tasks.testing.logging.TestLogEvent
 import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
 
@@ -15,45 +14,6 @@ buildscript {
 
 plugins {
     alias(kotlinx.plugins.serialization) apply false
-}
-
-subprojects {
-    tasks.withType<KotlinJvmCompile> {
-        kotlinOptions {
-            jvmTarget = JavaVersion.VERSION_17.toString()
-        }
-    }
-
-    tasks.withType<Test> {
-        useJUnitPlatform()
-        testLogging {
-            events(TestLogEvent.PASSED, TestLogEvent.SKIPPED, TestLogEvent.FAILED)
-        }
-    }
-
-    plugins.withType<BasePlugin> {
-        plugins.apply("detekt")
-        configure<BaseExtension> {
-            compileSdkVersion(AndroidConfig.compileSdk)
-            defaultConfig {
-                minSdk = AndroidConfig.minSdk
-                targetSdk = AndroidConfig.targetSdk
-                ndk {
-                    version = AndroidConfig.ndk
-                }
-            }
-
-            compileOptions {
-                sourceCompatibility = JavaVersion.VERSION_17
-                targetCompatibility = JavaVersion.VERSION_17
-                isCoreLibraryDesugaringEnabled = true
-            }
-
-            dependencies {
-                add("coreLibraryDesugaring", libs.desugar)
-            }
-        }
-    }
 }
 
 tasks.register<Delete>("clean") {
