+++ buildSrc/src/main/kotlin/AndroidConfig.kt
@@ -3,4 +3,8 @@ object AndroidConfig {
     const val minSdk = 26
     const val targetSdk = 34
     const val ndk = "26.1.10909125"
+    const val namespace = "eu.kanade.tachiyomi"
+    const val applicationId = "app.mihon"
+    const val versionCode = 3
+    const val versionName = "0.16.2"
 }
