+++ data/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
-    id("com.android.library")
-    kotlin("android")
+    id("android-module-setup")
     kotlin("plugin.serialization")
     id("app.cash.sqldelight")
 }
